refactor: remove transaction lib on cluster code dependency (#4417)
